TST: test_plots.py reorganization.
