+++ tests/unit/test_plots.py
@@ -1,0 +1,42 @@
+import os
+from unittest.mock import patch
+
+import matplotlib.pyplot as plt
+
+from rocketpy import Flight
+from rocketpy.plots.compare import Compare, CompareFlights
+
+
+@patch("matplotlib.pyplot.show")
+def test_compare(mock_show, flight_calisto):
+    """Here we want to test the 'x_attributes' argument, which is the only one
+    that is not tested in the other tests.
+
+    Parameters
+    ----------
+    mock_show :
+        Mocks the matplotlib.pyplot.show() function to avoid showing the plots.
+    flight_calisto : rocketpy.Flight
+        Flight object to be used in the tests. See conftest.py for more details.
+    """
+    flight = flight_calisto
+
+    objects = [flight, flight, flight]
+
+    comparison = Compare(object_list=objects)
+
+    fig, _ = comparison.create_comparison_figure(
+        y_attributes=["z"],
+        n_rows=1,
+        n_cols=1,
+        figsize=(10, 10),
+        legend=False,
+        title="Test",
+        x_labels=["Time (s)"],
+        y_labels=["Altitude (m)"],
+        x_lim=(0, 3),
+        y_lim=(0, 1000),
+        x_attributes=["time"],
+    )
+
+    assert isinstance(fig, plt.Figure) == True
